Changed serializers to allow floats

+++ weather/api/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 class WeatherSerializer(serializers.Serializer):
-    maximum = serializers.IntegerField()
-    minimum = serializers.IntegerField()
-    average = serializers.IntegerField()
-    median = serializers.IntegerField()
+    maximum = serializers.FloatField()
+    minimum = serializers.FloatField()
+    average = serializers.FloatField()
+    median = serializers.FloatField()
